refactor(api): move ComplementaryCertificationHabilitation from model to models
