+++ api/tests/organizational-entities/unit/domain/models/ComplementaryCertificationHabilitation_test.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { ComplementaryCertificationHabilitation } from '../../../../../src/organizational-entities/domain/model/ComplementaryCertificationHabilitation.js';
+import { ComplementaryCertificationHabilitation } from '../../../../../src/organizational-entities/domain/models/ComplementaryCertificationHabilitation.js';
 import { expect } from '../../../../test-helper.js';
 import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
